+++ pkg/configparser/configparser.go
@@ -15,9 +15,18 @@ var (
 
 // Source acts as a generic type for different source of configs.
 // ex: env, yaml, toml. refers to EnvSource for envars parsing.
+// Source also have the ability to marshall and unmarshal given config to file.
 type Source interface {
+
+	// GetValue will return our requested value from key.
+	// This will throw error if none was found or given key doesn't
+	// follow our regex parsing.
 	GetValue(key string) (interface{}, error)
+
+	// Name returns name of given source.
 	Name() string
+
+	// Marshal serializes a data stream to an io.Writer from
 }
 
 // ConfigManager holds types for generic managers to generate configs.
@@ -58,75 +67,7 @@ func (c *ConfigManager) Load() {
 	}
 }
 
-// Options is our variable configuration.
-type Options struct {
-	// Name will have format iris.option1.option2
-	Name         string
-	Description  string
-	DefaultValue interface{}
-	LoadedValue  interface{}
-	Manager      *ConfigManager
-	ConfigSource Source
-}
-
-// LoadValue will load given values if exists, otherwise use default ones.
-func (o *Options) LoadValue() {
-	def := o.DefaultValue
-	o.ConfigSource = nil
-
-	for i := len(o.Manager.sources) - 1; i >= 0; i-- {
-		source := o.Manager.sources[i]
-		// v would be value from given source, check envsource.go for examples
-		v, _ := source.GetValue(o.Name)
-
-		if v != nil {
-			def = v
-			o.ConfigSource = source
-			break
-		}
-	}
-
-	if o.DefaultValue != nil {
-		if _, ok := o.DefaultValue.(int); ok {
-			def = interface{}(toIntVal(def))
-		} else if _, ok = o.DefaultValue.(bool); ok {
-			def = interface{}(toBoolVal(def))
-		}
-	}
-
-	o.LoadedValue = def
-}
-
-// UpdateValue updates loaded value.
-func (o *Options) UpdateValue(val interface{}) {
-	switch val.(type) {
-	case bool:
-		o.LoadedValue = toBoolVal(val)
-	case string:
-		o.LoadedValue = toStrVal(val)
-	case int:
-		o.LoadedValue = toIntVal(val)
-	case float64:
-		o.LoadedValue = toFloat64Val(val)
-	}
-}
-
-// GetString are a getter string for &Options.LoadedValue.
-func (o *Options) GetString() string {
-	return toStrVal(o.LoadedValue)
-}
-
-// GetInt are a getter int for &Options.LoadedValue.
-func (o *Options) GetInt() int {
-	return toIntVal(o.LoadedValue)
-}
-
-// GetBool are a getter bool for &Options.LoadedValue.
-func (o *Options) GetBool() bool {
-	return toBoolVal(o.LoadedValue)
-}
-
-// GetFloat are a getter float64 for &Options.LoadedValue.
-func (o *Options) GetFloat() float64 {
-	return toFloat64Val(o.LoadedValue)
+func (c *ConfigManager) Clear() {
+	c.sources = make([]Source, 0)
+	c.Options = make(map[string]*Options)
 }
